Added browser.html
This file acts as the "catcher" in the target environment. Previously I relied on notepad or another text editor. However, that is not particularly stealthy. Using the developer tools in IE/Edge is not terribly suspicious. Then this takes care of base64 decoding without needing certutil.exe, or powershell.exe or other more suspicious ways to do it.